rename: 카드 입력 정보 context API 파일명 변경
