+++ src/components/registerForm/securityCode/index.tsx
@@ -4,7 +4,7 @@ import FormLabel from "src/components/@common/FormLabel";
 import Input from "src/components/@common/Input";
 import { ONLY_NUMBER_REGEXP } from "src/utils/regexp";
 import styled, { css } from "styled-components";
-import { InputValuesContext } from "../Main";
+import { InputValuesContext } from "../InputValueContext";
 
 function SecurityCode() {
   const [cardInput, setCardInput] = useContext(InputValuesContext);
